Add ratings node to root
"ratings" describes how to apply ratings to paths that have issues. For
now, we only support a "paths" sub-node which is a list of paths or
patterns identifying which files to apply ratings to when issues are
present.

If no ratings.paths are present, no files will be rated.

Refactorings:

- Extracted Glob and GlobList to reuse between exclude and ratings paths
- Moved tests of Sequence behavior to a Sequence unit spec
- Consistently used minitest/spec syntax and double quotes