Allow engine config to have arbitrary mappings and depths
Previously, we've been explicitly mapping certain keys to certain types of
nodes. However, we now need the ability to parse arbitrary configurations under
the engine's `config` key, and the configurations could have arbitrary depths.

To do this, we must explicitly check the type of the values that are found in
the configurations and then map them to their appropriate node types.
`NestedConfig` is introduced to handle this recursive behavior.