Don't emit "no keys" error when other errors occurred
Invalid mappings are removed from the result during the parsing phase,
which leads to somewhat confusing error messages when you have an
invalid engines key: first you're notified your engines key is invalid,
and then you're notified that your config is invalid because it doesn't
have an engine config.

So we should only report this error if we don't have other errors to
report.