Add engine issue_override support
In preparation for a new CC CLI feature the config needs to be expanded.

CC CLI will soon be able to override severity of issues emitted by the
engine.

One use case is lowering engine to all-info issues. This is useful when
new engines are introduced to a big old codebase. One would want to see
issues to be able to fix them but they won't block ongoing work. The
effort to eliminate all the issues at once might be substantial. It's
much more tenable to gradually fix them as the regular work continues.